[SD-1723] A combination of privileges gives unclear error
Modified SuperdeskAPIError.__str__() in errors.py module to return error code and message, which is sometimes sent to client as part of response. This will enhance the user experience to see what exactly went wrong. To put it in simple words the error message is changed from "Error:403" to "Error: 403: Unauthorized to modify unique name".