fix(workspace): prevent preferences update errors
there were multiple calls to update preferences at the same time,
both using same etag, so the later had to fail. that was happening
on desk/stage change which are 2 settings. now those 2 should be
updated together, and also there is no update if the new value
matches the old one.

SD-1867