[Scanpix] Simplified rendition images selection
the former way of choosing rendition images was to do a best guest
based on actual rendition sizes in settings, and it was not settings
correclty all renditions keys.

As renditions sizes have little chance to change, the code was
simplified and Superdesk renditions to Scanpix preview mapping is now
done with a simple dict containing a list of preview to get in order of
preference.

SDNTB-210