[NITF Formatter] fixed featuremedia and body_html handling
only "featureimage" was used, now "featuremedia" is checked too if
"featureimage" is not found

body_html was added to NITF's <body.content/> using the inherited
method from core's NITF Formatter, but this was not getting all the
elements.
To work around this, the inherited method is not used anymore, and
BeautifulSoup is used to transform the HTML to XML, which is then added
to body.content

This commit fixes the issues reported in feedback of SDNTB-229 (in the
comments)

formatter tests have also been improved, and a fallback has been added
if original rendition doesn't exist.